requirements.txt
asgiref==3.9.1
click==8.3.1
colorama==0.4.6
Django==5.2.6
django-ckeditor-5==0.2.18
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.1
psycopg2-binary==2.9.10
pytokens==0.3.0
reportlab==4.0.9
sqlparse==0.5.3
tzdata==2025.2
